Fix crash due to lsadrv-build-3.10.patch
The patch reused a procfs-specific hook as read hook in a file_operations
struct, however they have a different signature. The lsadrv_read_devices hook
would hence be called with incorrect parameters, resulting in a kernel crash.

Instead, we rewrite the procfs entry code to use the single_open seq_file hook,
which also happens to be simpler and safer.