Add support for 32bit ioctl() codes
Because some some members of the structs used as
lsadrv_usb_ioctl args don't have a fixed size,
the ioctl codes differ and are not catched by the switch
when built on 64bit.
We can't just fix the struct declaration since some members are pointers, and changing to some addr_t would likely break the userland code we can't change.
We declare them separately along with the 32bit-fixed structs,
and call ourselves back with the fixed 64bit struct and codes.

Note because we are being called through devio, we can't provide
a compat_ioctl() call instead, so we do it inline.